requirements.txt
Django==1.11.5
django-crudbuilder==0.2.5
django-tables2==1.13.0